Zmiana typu zmiennej stanu uint16_t -> uint8_t

+++ dynabox.cpp
@@ -14,7 +14,7 @@ Dynabox::Dynabox() : Machine(ST_MAX_STATES)
 
 }
 
-uint16_t Dynabox::StartupTest()
+uint8_t Dynabox::StartupTest()
 {
 	motor.EV_Homing();
 	timer.Assign(TIMER_MODBUS_RTU_POLL, 50, ModbusPoll);
